+++ src/Sentry/ISentryScopeManager.cs
@@ -53,5 +53,7 @@ public interface ISentryScopeManager
     /// </remarks>
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
+    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
+              "that provide a callback to a configurable scope.")]
     void WithScope(Action<Scope> scopeCallback);
 }