+++ src/Sentry/SentrySdk.cs
@@ -331,6 +331,8 @@ public static partial class SentrySdk
     /// </remarks>
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
+    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
+              "that provide a callback to a configurable scope.")]
     [DebuggerStepThrough]
     public static void WithScope(Action<Scope> scopeCallback)
         => CurrentHub.WithScope(scopeCallback);
@@ -346,6 +348,8 @@ public static partial class SentrySdk
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
     /// <returns>The result from the callback.</returns>
+    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
+              "that provide a callback to a configurable scope.")]
     [DebuggerStepThrough]
     public static T? WithScope<T>(Func<Scope, T?> scopeCallback)
         => CurrentHub is IHubEx hub ? hub.WithScope(scopeCallback) : default;
@@ -362,6 +366,8 @@ public static partial class SentrySdk
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
     /// <returns>An async task to await the callback.</returns>
+    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
+              "that provide a callback to a configurable scope.")]
     [DebuggerStepThrough]
     public static Task WithScopeAsync(Func<Scope, Task> scopeCallback)
         => CurrentHub is IHubEx hub ? hub.WithScopeAsync(scopeCallback) : Task.CompletedTask;
@@ -378,6 +384,8 @@ public static partial class SentrySdk
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
     /// <returns>An async task to await the result of the callback.</returns>
+    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
+              "that provide a callback to a configurable scope.")]
     [DebuggerStepThrough]
     public static Task<T?> WithScopeAsync<T>(Func<Scope, Task<T?>> scopeCallback)
         => CurrentHub is IHubEx hub ? hub.WithScopeAsync(scopeCallback) : Task.FromResult(default(T));