+++ test/Sentry.Tests/SentrySdkTests.cs
@@ -469,6 +469,7 @@ public class SentrySdkTests : IDisposable
         }
     }
 
+    [Obsolete]
     [Fact]
     public void WithScope_DisabledSdk_CallbackNeverInvoked()
     {
@@ -479,6 +480,7 @@ public class SentrySdkTests : IDisposable
         Assert.False(invoked);
     }
 
+    [Obsolete]
     [Fact]
     public void WithScopeT_DisabledSdk_CallbackNeverInvoked()
     {
@@ -486,6 +488,7 @@ public class SentrySdkTests : IDisposable
         Assert.False(invoked);
     }
 
+    [Obsolete]
     [Fact]
     public async Task WithScopeAsync_DisabledSdk_CallbackNeverInvoked()
     {
@@ -498,6 +501,7 @@ public class SentrySdkTests : IDisposable
         Assert.False(invoked);
     }
 
+    [Obsolete]
     [Fact]
     public async Task WithScopeAsyncT_DisabledSdk_CallbackNeverInvoked()
     {
@@ -505,6 +509,7 @@ public class SentrySdkTests : IDisposable
         Assert.False(invoked);
     }
 
+    [Obsolete]
     [Fact]
     public void WithScope_InvokedWithNewScope()
     {
@@ -531,6 +536,7 @@ public class SentrySdkTests : IDisposable
         SentrySdk.ConfigureScope(s => Assert.Same(expected, s));
     }
 
+    [Obsolete]
     [Fact]
     public void WithScopeT_InvokedWithNewScope()
     {
@@ -555,6 +561,7 @@ public class SentrySdkTests : IDisposable
         SentrySdk.ConfigureScope(s => Assert.Same(expected, s));
     }
 
+    [Obsolete]
     [Fact]
     public async Task WithScopeAsync_InvokedWithNewScope()
     {
@@ -584,6 +591,7 @@ public class SentrySdkTests : IDisposable
         SentrySdk.ConfigureScope(s => Assert.Same(expected, s));
     }
 
+    [Obsolete]
     [Fact]
     public async Task WithScopeAsyncT_InvokedWithNewScope()
     {